IT PRINTS MY DUDES

+++ support/support.c
@@ -1,14 +1,15 @@
 #include <stdio.h>
+#include <sys/types.h>
 
 void print(int);
-int read(void);
+int64_t read(void);
 
 
 void print(int arg) {
         printf("%d\n", arg);
 }
 
-int read(void) {
+int64_t read(void) {
         int arg;
         scanf("%d", &arg);
         return arg;
